fix(content): import zod-to-json-schema only for zod v3 schemas
