+++ src/utils/index.ts
@@ -6,4 +6,4 @@ export { defineTransformer } from './content/transformers/utils'
 /**
  * This is only for backward compatibility with Zod v3. Consider using direct import from 'zod' instead. This will be removed in the next major version.
  */
-export { z } from './schema/zod3'
+export { z } from './schema/zod3-legacy'
